MLE-12287 Now supporting all connection params
Going to do some manual testing of this with a 2-way SSL setup, but we expect these all to work as they get passed down to the Java Client which knows how to construct the appropriate client.

Re: the `ConnectionInputs` class - I am going to do a refactor soon to make a `WriteInputs` class to avoid some of the duplication between WriteDocumentParams and CopyCommand.