First cut of documentation
This is baby step #1. Includes an example project. 

Want to get this merged and will then submit a follow-up PR that fixes typos and makes general improvements for an internal employee to be able to consume this. 

Made a few tweaks to logging and command descriptions too. Added a new gzip test for CSV as I wasn't 100% sure if it works (it does).