ML-12823 Can now import JSON array files
Reworked the command for JSON lines to be for JSON files in general, defaulting to files containing JSON objects and arrays of objects.

If a user has JSON lines files, they simply include `--jsonLines` as an option. 

Also noted the gap we have here with not being able to use the URI template feature when the JSON files are in a zip. Will address that in a future story.