MLE-12866 uriTemplate no longer appears for import_rdf_files
This was a bit awkward to do, as it required removing `uriTemplate` from `WriteDocumentParams`. So I made a subclass - `WriteDocumentWithTemplateParams` - whose name I'm not crazy about yet. 

Also made `ReadFilesParams` so that the command for importing RDF files and the abstract command for importing any kind of files don't have any duplication.