MLE-13922 Can now depend on NT as a library
The Gradle publishing config is in line with the latest in the Gradle docs. It's simpler than what we have in our other projects, so an unrelated cleanup effort will be to "modernize" what we have for Gradle publishing in all of our projects. 

Fixed all the Javadoc warnings too, just so they don't show up and we don't have to suppress them.