Added interfaces for aggregate XML, archives, Avro, and MLCP archives
Lot of little changes here, as I made `AbstractCommand` implement `Executor`, which forced every command class to be updated to declare a type associated with `Executor`. 

Still no new functionality, just exposing stuff in the API interfaces. 

Will add Javadocs to all the API interfaces before the release. 

Also got rid of `WriteDocumentWithTemplateParams`, in favor of letting it exist for importing RDF. It may work soon - i.e. once we support XPath expressions in the template, which is doable.