Repackaging - created `cli` and `impl` packages
No functional changes here - just moving a bunch of classes around. We now have 3 packages - `api`, `cli`, and `impl`. Tests are updated based on these packages as well.

We could break up the `api` package at some point, but given that it's only interfaces and enums, it's fairly simple to understand for now.