Refactored how the preview feature works
I was looking into enhancing this to support showing the dataset schema as well and realized that the current impl is a bit messy. This cleans things up.

Also, the feature of "Don't require a connection string when doing a preview" was very half-baked - only worked for imports - and so I removed it for now until it can be properly supported.