Tweaking JSON import docs
The naming here is tricky because regular JSON object files are handled by `import-files`, and this is for two different types of files that need special handling - JSON Lines and arrays of JSON objects, each of which produce multiple documents in MarkLogic. `import-json-files` doesn't quite capture that the intent is really more of "import special JSON files".