MLE-19164 Configuring Spark master URL based on partition count
This is for export operations using Optic, where the user is able to specify the number of partitions.

Updated the docs as well, though I'm thinking the docs could use a scrub soon to iron out the subtleties between "partitions" and "Spark threads" and "partition threads".

And - fixed a gitignore issue that had hidden a couple of export tests. Odd.