MLE-21745 Refactor: Improved TdeBuilder interface
This doesn't complete the story, but I realized it would help a lot if TdeBuilder returns an AbstractWriteHandle, which easily plugs into an eval call.

There's a big TODO on making an actual connection to load the TDE. This PR is just to get the refactor of TdeBuilder in place.