MLE-21726 Refactor - switching from List to Map for var params
I think I did a List initially so I could provide custom error messages. But the default error messages from picocli for mismatched key/value pairs with a Map param are good enough - they identify the bad param, which is enough for the user to fix the problem. So this gets rid of some code.