MLE-22714 Can now specify document type when importing an archive
Had to do some deprecation in the API layer in order to introduce a new interface with a method for setting the document type. Had a discussion with Copilot on what to call the new "to" method. Settled on "toOptions" for now, which isn't great but we'll do Flux 2.0 soon enough to clean it up.