MLE-23146 Fixing compiler warnings
Per Copilot, some of the Params class need to suppress warnings as there's no way around e.g. returning "T".

Added `IReadFilesParams` to allow for a number of unchecked warnings to be removed. Going to rename that in a follow up PR.

Also removed usage of wildcards per Copilot's advice, and using method-level generics instead. That likewise removes warnings that otherwise are forced onto the user.