MLE-23024 Running on Spark 4
Lot of little changes, but overhaul, a simpler set of dependencies as we don't have to jump through as many hoops to avoid transitive vulnerabilities.

This will likely have a test failure due to the deprecated fileRows.documentType option, as the pipeline for that PR is stuck right now. Will ignore that test failure for now.