MLE-23618 Added new Main method to return a Command object
This is for ragplus, so that a custom Spark session - specifically with a Spark listener - can be used to execute the command.

This also ended up being a better approach for our "options tests" that just need to parse options but not execute the command. So about a dozen tests were modified to use this approach. They also needed a connection string added since the command is being validated now too.