MLE-26348 Reworking docs on parallelism and partitions
This also removes a feature from 1.2.0 that was misguided, where the Spark master URL was overridden based on partition count. The master URL defines the number of cores available for processing partitions in parallel. Setting it to the number of partitions wasn't going to be harmful - it was still limited by how many cores are available - but it's misleading to the user. That feature and the associated docs / tests have been removed as well.

Generally, `local[*]` is going to be ideal for all use cases, where Spark uses all available cores. An advanced user does not need docs on how to change that because they'll have a keen understanding of the Spark master URL already.